refactor(app): the dungeons value maps and the behaviour they dispatch to
constant-maps puts a value in services/, and every loader map plus the
battle StateMap sat in models/. The battle states had to follow the map
that dispatches to them — they drive stores and scene tweens rather than
describing a shape. The Controls class hierarchy stays a model, like the
sheet command classes it mirrors. ExtraTilesetKey stops being exported by
a file nothing outside reads it from.

Closes app/models/dungeons; app/models splits into the areas left open.
