+++ packages/app/app/services/dungeons/loader/image/InventoryLoaderMap.ts
