+++ packages/app/app/services/dungeons/loader/image/MonsterPartyLoaderMap.ts
