+++ packages/app/app/services/dungeons/loader/image/UI/MonsterLoaderMap.ts
