+++ packages/app/app/services/dungeons/loader/sound/SoundLoaderMap.ts
@@ -1,5 +1,5 @@
-import { BackgroundMusicLoaderMap } from "@/models/dungeons/loader/sound/BackgroundMusicLoaderMap";
-import { SoundEffectLoaderMap } from "@/models/dungeons/loader/sound/SoundEffectLoaderMap";
+import { BackgroundMusicLoaderMap } from "@/services/dungeons/loader/sound/BackgroundMusicLoaderMap";
+import { SoundEffectLoaderMap } from "@/services/dungeons/loader/sound/SoundEffectLoaderMap";
 import { mergeObjectsStrict } from "@esposter/shared";
 
 export const SoundLoaderMap = mergeObjectsStrict(BackgroundMusicLoaderMap, SoundEffectLoaderMap);
