+++ packages/app/app/services/dungeons/loader/spritesheet/SpritesheetLoaderMap.ts
@@ -1,5 +1,5 @@
-import { AttackLoaderMap } from "@/models/dungeons/loader/spritesheet/AttackLoaderMap";
-import { CharacterLoaderMap } from "@/models/dungeons/loader/spritesheet/CharacterLoaderMap";
+import { AttackLoaderMap } from "@/services/dungeons/loader/spritesheet/AttackLoaderMap";
+import { CharacterLoaderMap } from "@/services/dungeons/loader/spritesheet/CharacterLoaderMap";
 import { mergeObjectsStrict } from "@esposter/shared";
 
 export const SpritesheetLoaderMap = mergeObjectsStrict(AttackLoaderMap, CharacterLoaderMap);
