+++ packages/app/app/services/dungeons/scene/battle/battleStateMachine.ts
@@ -1,4 +1,4 @@
-import { StateMap } from "@/models/dungeons/state/battle/StateMap";
-import { StateMachine } from "@/models/dungeons/state/StateMachine";
+import { StateMap } from "@/services/dungeons/state/battle/StateMap";
+import { StateMachine } from "@/services/dungeons/state/StateMachine";
 
 export const battleStateMachine = new StateMachine(StateMap);
